Slides services duplicated file eliminated, all services import fixed and
