+++ src/Components/Backoffice/Home/EditHomeForm.js
@@ -2,7 +2,7 @@ import React, { useState, useEffect } from "react";
 import { useFormik } from "formik";
 import {Button} from "@mui/material";
 import EditHomeItem from "./EditHomeItem";
-import { getSlides } from "../../../Services/slidesService";
+import { getSlides } from "../../../Services/slidesServices";
 import {Table,TableBody,TableCell,TableContainer,TableHead,TableRow} from "@mui/material";
 import "../../FormStyles.css";
 import Swal from "sweetalert2";
