+++ src/Components/Carousel/Carousel.js
@@ -1,5 +1,5 @@
 import React, { useEffect, useState } from "react";
-import { getImagesSlides } from "../../Services/slidesService";
+import { getImagesSlides } from "../../Services/slidesServices";
 import { Swiper, SwiperSlide } from "swiper/react";
 import SwiperCore, {
   Autoplay,
